Text related logic moved to a separate TextBehavior
